Expect the timestamp to be an integer, not just a float

+++ repoze/postoffice/tests/test_api.py
@@ -345,7 +345,7 @@ class TestAPI(unittest.TestCase):
         self.assertEqual(len(A), 1)
         m = A.pop_next()
         self.assertEqual(m, 'one')
-        self.assertTrue(float(m['X-Postoffice-Date']))
+        self.assertTrue(int(m['X-Postoffice-Date']))
         B = queues['B']
         self.assertEqual(len(B), 2)
         self.assertEqual(B.pop_next(), 'three')
